bot/requirements.txt
aiogram==3.*
dnspython==2.*
cryptography==42.*
aiohttp==3.*
python-dotenv==1.*
python-whois==0.9.*
pytest==8.*
pytest-asyncio==0.23.*
uvloop>=0.19
idna>=3.0